Add vim-font-size
See https://github.com/changa/vim-fontsize

Fontsize provides key mappings for adjusting the font size in Gvim. The initial
mappings launch a "font size" mode in which single keypresses are used to
adjust the font. This mode ends after a timeout or when the user presses any
key not defined in the mode.

Mappings:
  * <leader><leader>+ increases font size
  * <leader><leader>- decreases font size
  * <leader><leader>0 resets font size